# API-SPECIFICATION.yml
openapi: 3.0.3
info:
  title: Payment Processing System API
  description: |
    API for payment flows (Authorize.Net sandbox integration). Contains endpoints for Purchase, Authorize, Capture, Cancel (void), Refund, Recurring Billing (Subscriptions), and Webhook handling. 
    JWT bearer is used to protect payment endpoints. Webhooks use HMAC signature validation. 
    All mutating endpoints support Idempotency-Key header for safe retries. 
    All responses include X-Correlation-ID for distributed tracing.
  version: 2.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /payments/subscriptions:
    post:
      summary: Create a recurring billing subscription
      tags:
        - subscriptions
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          description: Validation error
        '502':
          description: Gateway error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentErrorResponse'
  /payments/subscriptions/{id}:
    get:
      summary: Get subscription by ID
      tags:
        - subscriptions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Subscription details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '404':
          description: Subscription not found
    put:
      summary: Update a subscription
      tags:
        - subscriptions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubscriptionRequest'
      responses:
        '200':
          description: Subscription updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '404':
          description: Subscription not found or update failed
    delete:
      summary: Cancel a subscription
      tags:
        - subscriptions
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Subscription cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  status:
                    type: string
                    example: cancelled
                  cancelled_at:
                    type: string
                    format: date-time
        '404':
          description: Subscription not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Card:
      type: object
      properties:
        number:
          type: string
          description: Card PAN
          example: "4111111111111111"
        expMonth:
          type: integer
          example: 12
        expYear:
          type: integer
          example: 2030
        cvv:
          type: string
          example: "123"
      required:
        - number
        - expMonth
        - expYear
        - cvv
...
    PaymentErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: GATEWAY_TIMEOUT
            message:
              type: string
              example: The payment gateway timed out
            category:
              type: string
              example: GATEWAY_ERROR
            retryable:
              type: boolean
              example: true
            retry_after_seconds:
              type: integer
              example: 5
            provider_error:
              type: object
              properties:
                code:
                  type: string
                message:
                  type: string
            suggestions:
              type: array
              items:
                type: string
            request_id:
              type: string
              example: req_abc123
            timestamp:
              type: string
              format: date-time
    CreateSubscriptionRequest:
      type: object
      required:
        - name
        - amount
        - startDate
        - card
      properties:
        name:
          type: string
          example: Monthly Premium Plan
        amount:
          type: number
          format: decimal
          example: 29.99
        intervalLength:
          type: integer
          example: 1
          default: 1
        intervalUnit:
          type: string
          enum: [days, months]
          example: months
          default: months
        startDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          example: "2026-04-01"
        card:
          $ref: '#/components/schemas/Card'
    UpdateSubscriptionRequest:
      type: object
      properties:
        name:
          type: string
          example: Annual Premium Plan
        amount:
          type: number
          format: decimal
          example: 49.99
    SubscriptionResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        gateway_subscription_id:
          type: string
          example: "12345"
        name:
          type: string
          example: Monthly Premium Plan
        amount:
          type: number
          example: 29.99
        interval:
          type: string
          example: "1 months"
        start_date:
          type: string
          example: "2026-04-01"
        status:
          type: string
          enum: [active, suspended, cancelled, expired, failed]
          example: active
        created_at:
          type: string
          format: date-time
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Idempotency key for safe retries (24h TTL). Same key + same body returns cached response.
      schema:
        type: string
        example: "purchase-order-123-v1"
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: Correlation ID for distributed tracing. Auto-generated if not provided.
      schema:
        type: string
        example: "abc12345-6789-def0"
...